+++ frontend/src/libs/firebase/fetchImageUrl.ts
@@ -1,5 +1,5 @@
 import { ref, getDownloadURL } from 'firebase/storage';
-import { storage } from '../../firebase'; // FIXME: 絶対パスで指定するようにする
+import { storage } from './firebase'; // FIXME: 絶対パスで指定するようにする
 
 export const fetchImageUrl = async (fileName: string) => {
   let url = '';
