+++ frontend/src/libs/firebase/uploadImage.ts
@@ -1,6 +1,6 @@
 import { ref, uploadBytes } from 'firebase/storage';
 import { generateRandomString } from '@/utils/generateRandomString';
-import { storage } from '../../firebase'; // FIXME: 絶対パスで指定するようにする
+import { storage } from './firebase'; // FIXME: 絶対パスで指定するようにする
 
 export const uploadImage = async (file: File) => {
   let fileName = generateRandomString(30);
